refactor: `lib/migrate.js` → `migrations/index.js`
