+++ test/db.test.js
@@ -1,5 +1,5 @@
 import pg from 'pg'
-import { migrate } from '../lib/migrate.js'
+import { migrate } from '../migrations/index.js'
 
 const { DATABASE_URL } = process.env
 
@@ -20,11 +20,14 @@ describe('spark-api database', () => {
   })
 
   it('allows multiple storage deals for the same CID', async () => {
+    const DUMMY_CID = 'bafyone'
+    await client.query('DELETE FROM retrievable_deals WHERE cid = $1', [DUMMY_CID])
+
     await client.query(`
       INSERT INTO retrievable_deals (cid, miner_id, expires_at)
       VALUES ($1, $2, $3), ($1, $4, $3)
     `, [
-      'bafyone',
+      DUMMY_CID,
       'f010',
       new Date(),
       'f020'
